// ui/src/lib/api/model/workPool.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BrowserGrid API
 * BrowserGrid is a distributed browser automation platform that provides scalable browser sessions and worker pool management.
 * OpenAPI spec version: 1.0
 */
import type { WorkPoolDefaultEnv } from './workPoolDefaultEnv';
import type { ProviderType } from './providerType';

/**
 * Work pool configuration for managing browser workers
 */
export interface WorkPool {
  auto_scale?: boolean;
  created_at?: string;
  default_env?: WorkPoolDefaultEnv;
  default_image?: string;
  default_priority?: number;
  description?: string;
  id?: string;
  max_concurrency?: number;
  max_idle_time?: number;
  max_session_duration?: number;
  min_size?: number;
  name?: string;
  paused?: boolean;
  provider?: ProviderType;
  queue_strategy?: string;
  updated_at?: string;
}