// ui/src/lib/api/model/providerType.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BrowserGrid API
 * BrowserGrid is a distributed browser automation platform that provides scalable browser sessions and worker pool management.
 * OpenAPI spec version: 1.0
 */

/**
 * Type of provider for a work pool
 */
export type ProviderType = typeof ProviderType[keyof typeof ProviderType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProviderType = {
  ProviderDocker: 'docker',
  ProviderACI: 'azure_aci',
  ProviderLocal: 'local',
} as const;